// Router-main/src/types.ts
export interface CidadeRota {
  id?: number;
  cidade: string; // e.g. "ALFENAS"
  alias: string;  // e.g. "ALFENAS-MG, ALFENA, RTA-ALFENAS"
  setor: string;  // e.g. "SUL-1"
  rota: string;   // e.g. "ROTA 06"
  prazo_padrao: number; // e.g. 2 for D+2
  prioridade_operacional: 'CRÍTICA' | 'ALTA' | 'NORMAL' | 'BAIXA';
  segunda?: boolean;
  terca?: boolean;
  quarta?: boolean;
  quinta?: boolean;
  sexta?: boolean;
  cod?: string;
}